# improve vector/glyph rendering quality using a sub-pixel rendering techinique