--- a/src/ImageSharp.Drawing/Brushes/PatternBrush{TColor}.cs
+++ b/src/ImageSharp.Drawing/Brushes/PatternBrush{TColor}.cs
@@ -47,12 +47,8 @@ namespace ImageSharp.Drawing.Brushes
         /// <summary>
         /// The pattern.
         /// </summary>
-        private readonly TColor[][] pattern;
-
-        /// <summary>
-        /// The stride width.
-        /// </summary>
-        private readonly int stride;
+        private readonly Fast2DArray<TColor> pattern;
+        private readonly Fast2DArray<Vector4> patternVector;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="PatternBrush{TColor}"/> class.
@@ -60,26 +56,23 @@ namespace ImageSharp.Drawing.Brushes
         /// <param name="foreColor">Color of the fore.</param>
         /// <param name="backColor">Color of the back.</param>
         /// <param name="pattern">The pattern.</param>
-        public PatternBrush(TColor foreColor, TColor backColor, bool[,] pattern)
+        public PatternBrush(TColor foreColor, TColor backColor, Fast2DArray<bool> pattern)
         {
-            this.stride = pattern.GetLength(1);
-
-            // Convert the multidimension array into a jagged one.
-            int height = pattern.GetLength(0);
-            this.pattern = new TColor[height][];
-            for (int x = 0; x < height; x++)
+            Vector4 foreColorVector = foreColor.ToVector4();
+            Vector4 backColorVector = backColor.ToVector4();
+            this.pattern = new Fast2DArray<TColor>(pattern.Width, pattern.Height);
+            this.patternVector = new Fast2DArray<Vector4>(pattern.Width, pattern.Height);
+            for (int i = 0; i < pattern.Data.Length; i++)
             {
-                this.pattern[x] = new TColor[this.stride];
-                for (int y = 0; y < this.stride; y++)
+                if (pattern.Data[i])
                 {
-                    if (pattern[x, y])
-                    {
-                        this.pattern[x][y] = foreColor;
-                    }
-                    else
-                    {
-                        this.pattern[x][y] = backColor;
-                    }
+                    this.pattern.Data[i] = foreColor;
+                    this.patternVector.Data[i] = foreColorVector;
+                }
+                else
+                {
+                    this.pattern.Data[i] = backColor;
+                    this.patternVector.Data[i] = backColorVector;
                 }
             }
         }
@@ -91,13 +84,12 @@ namespace ImageSharp.Drawing.Brushes
         internal PatternBrush(PatternBrush<TColor> brush)
         {
             this.pattern = brush.pattern;
-            this.stride = brush.stride;
         }
 
         /// <inheritdoc />
         public BrushApplicator<TColor> CreateApplicator(PixelAccessor<TColor> sourcePixels, RectangleF region)
         {
-            return new PatternBrushApplicator(this.pattern, this.stride);
+            return new PatternBrushApplicator(sourcePixels, this.pattern, this.patternVector);
         }
 
         /// <summary>
@@ -106,30 +98,22 @@ namespace ImageSharp.Drawing.Brushes
         private class PatternBrushApplicator : BrushApplicator<TColor>
         {
             /// <summary>
-            /// The patter x-length.
-            /// </summary>
-            private readonly int xLength;
-
-            /// <summary>
-            /// The stride width.
-            /// </summary>
-            private readonly int stride;
-
-            /// <summary>
             /// The pattern.
             /// </summary>
-            private readonly TColor[][] pattern;
+            private readonly Fast2DArray<TColor> pattern;
+            private readonly Fast2DArray<Vector4> patternVector;
 
             /// <summary>
             /// Initializes a new instance of the <see cref="PatternBrushApplicator" /> class.
             /// </summary>
+            /// <param name="sourcePixels">The sourcePixels.</param>
             /// <param name="pattern">The pattern.</param>
-            /// <param name="stride">The stride.</param>
-            public PatternBrushApplicator(TColor[][] pattern, int stride)
+            /// <param name="patternVector">The patternVector.</param>
+            public PatternBrushApplicator(PixelAccessor<TColor> sourcePixels, Fast2DArray<TColor> pattern, Fast2DArray<Vector4> patternVector)
+            : base(sourcePixels)
             {
                 this.pattern = pattern;
-                this.xLength = pattern.Length;
-                this.stride = stride;
+                this.patternVector = patternVector;
             }
 
             /// <summary>
@@ -140,14 +124,14 @@ namespace ImageSharp.Drawing.Brushes
             /// <returns>
             /// The Color.
             /// </returns>
-            public override TColor this[int x, int y]
+            internal override TColor this[int x, int y]
             {
                 get
                 {
-                    x = x % this.xLength;
-                    y = y % this.stride;
+                    x = x % this.pattern.Height;
+                    y = y % this.pattern.Width;
 
-                    return this.pattern[x][y];
+                    return this.pattern[x, y];
                 }
             }
 
@@ -155,6 +139,33 @@ namespace ImageSharp.Drawing.Brushes
             public override void Dispose()
             {
                 // noop
+            }
+
+            internal override void Apply(float[] scanline, int scanlineWidth, int offset, int x, int y)
+            {
+                using (PinnedBuffer<float> buffer = new PinnedBuffer<float>(scanline))
+                {
+                    var slice = buffer.Slice(offset);
+
+                    for (var xPos = 0; xPos < scanlineWidth; xPos++)
+                    {
+                        int targetX = xPos + x;
+                        int targetY = y;
+
+                        float opacity = slice[xPos];
+                        if (opacity > Constants.Epsilon)
+                        {
+                            Vector4 backgroundVector = this.Target[targetX, targetY].ToVector4();
+                            Vector4 sourceVector = this.patternVector[targetX % this.pattern.Height, targetX % this.pattern.Width];
+
+                            Vector4 finalColor = Vector4BlendTransforms.PremultipliedLerp(backgroundVector, sourceVector, opacity);
+
+                            TColor packed = default(TColor);
+                            packed.PackFromVector4(finalColor);
+                            this.Target[targetX, targetY] = packed;
+                        }
+                    }
+                }
             }
         }
     }