--- a/src/ImageSharp.Drawing/Brushes/RecolorBrush{TColor}.cs
+++ b/src/ImageSharp.Drawing/Brushes/RecolorBrush{TColor}.cs
@@ -140,9 +140,9 @@ namespace ImageSharp.Drawing.Brushes
             {
                 using (PinnedBuffer<float> buffer = new PinnedBuffer<float>(scanline))
                 {
-                    var slice = buffer.Slice(offset);
+                    BufferPointer<float> slice = buffer.Slice(offset);
 
-                    for (var xPos = 0; xPos < scanlineWidth; xPos++)
+                    for (int xPos = 0; xPos < scanlineWidth; xPos++)
                     {
                         int targetX = xPos + x;
                         int targetY = y;