… src/ImageSharp.Drawing/Brushes/SolidBrush{TColor}.cs
… src/ImageSharp.Drawing/GraphicsOptions.cs
… src/ImageSharp.Drawing/ImageSharp.Drawing.csproj
… src/ImageSharp.Drawing/Paths/ShapeRegion.cs
… src/ImageSharp.Drawing/Processors/FillRegionProcessor.cs
… src/ImageSharp.Drawing/Region.cs
… src/ImageSharp.Drawing/Text/TextGraphicsOptions.cs
… src/ImageSharp/Common/Memory/BufferPointer{T}.cs
… tests/ImageSharp.Benchmarks/Drawing/FillPolygon.cs
… tests/ImageSharp.Tests/Drawing/FillPatternTests.cs
… tests/ImageSharp.Tests/Drawing/FillRegionProcessorTests.cs
… tests/ImageSharp.Tests/Drawing/Paths/ShapeRegionTests.cs
… tests/ImageSharp.Tests/Drawing/SolidBezierTests.cs
… tests/ImageSharp.Tests/Drawing/SolidComplexPolygonTests.cs
… tests/ImageSharp.Tests/Drawing/SolidPolygonTests.cs
… tests/ImageSharp.Tests/Drawing/Text/TextGraphicsOptionsTests.cs