### Prerequisites

- [x] I have written a descriptive pull-request title
- [x] I have verified that there are no overlapping [pull-requests](https://github.com/JimBobSquarePants/ImageSharp/pulls) open
- [x] I have verified that I am following matches the existing coding patterns and practise as demonstrated in the repository. These follow strict Stylecop rules :cop:.
- [x] I have provided test coverage for my change (where applicable)

### Description
Swap out the fill region processor to use a much more accurate sub-pixel rendering technique.

This new renderer is a tiny bit slower than the old one, but (and its a big but) it totally removes the Parrallel looping making it entirely single threaded.

#### New Sub-pixel renderer
![new](https://cloud.githubusercontent.com/assets/166440/24059207/0a73df04-0b46-11e7-8de2-26c7000a0163.png)
``` ini

BenchmarkDotNet=v0.10.1, OS=Microsoft Windows NT 6.2.9200.0
Processor=Intel(R) Core(TM) i5-6200U CPU 2.30GHz, ProcessorCount=4
Frequency=2343752 Hz, Resolution=426.6663 ns, Timer=TSC
  [Host]     : Clr 4.0.30319.42000, 64bit RyuJIT-v4.6.1637.0 [AttachedDebugger]
  DefaultJob : Clr 4.0.30319.42000, 64bit RyuJIT-v4.6.1637.0


```
|                        Method |       Mean |    StdErr |    StdDev | Scaled | Scaled-StdDev |     Gen 0 |    Gen 1 |    Gen 2 | Allocated |
|------------------------------ |----------- |---------- |---------- |------- |-------------- |---------- |--------- |--------- |---------- |
| 'System.Drawing Fill Polygon' |  4.0584 ms | 0.0485 ms | 0.1879 ms |   1.00 |          0.00 | 2069.5043 | 428.8793 | 428.8793 |   5.15 MB |
|     'ImageSharp Fill Polygon' | 13.4327 ms | 0.1318 ms | 0.7220 ms |   3.32 |          0.22 |  512.5000 | 262.5000 | 262.5000 |    5.3 MB |


#### Old renderer
![old](https://cloud.githubusercontent.com/assets/166440/24059208/0a74a948-0b46-11e7-87ab-188c3a8ff937.png)
``` ini

BenchmarkDotNet=v0.10.1, OS=Microsoft Windows NT 6.2.9200.0
Processor=Intel(R) Core(TM) i5-6200U CPU 2.30GHz, ProcessorCount=4
Frequency=2343752 Hz, Resolution=426.6663 ns, Timer=TSC
  [Host]     : Clr 4.0.30319.42000, 64bit RyuJIT-v4.6.1637.0 [AttachedDebugger]
  DefaultJob : Clr 4.0.30319.42000, 64bit RyuJIT-v4.6.1637.0


```
|                        Method |      Mean |    StdErr |    StdDev | Scaled | Scaled-StdDev |     Gen 0 |    Gen 1 |    Gen 2 | Allocated |
|------------------------------ |---------- |---------- |---------- |------- |-------------- |---------- |--------- |--------- |---------- |
| 'System.Drawing Fill Polygon' | 4.0581 ms | 0.0440 ms | 0.2108 ms |   1.00 |          0.00 | 2003.1250 | 362.5000 | 362.5000 |   5.15 MB |
|     'ImageSharp Fill Polygon' | 9.2461 ms | 0.0911 ms | 0.4988 ms |   2.28 |          0.16 |  765.6250 | 640.6250 | 640.6250 |   5.11 MB |