@JimBobSquarePants 2017-03-18
Is anyone else concerned about the sheer amount of GC in the benchmarks? I haven't done any investigation at all but the numbers seem very high to me.