@tocsoft 2017-03-18
I suspect the allocations are in the SixLabors.Shapes side of the process, this can probably be elevated by caching the shape objects and asking them to be filled directly instead of using the helper methods that take points and build the shape each time.