@antonfirsov 2017-03-18
Managed to track down allocations with the basic VS memory profiler! :)
The most allocations are done by 'SixLabors.Shapes.InternalPath.FindIntersections()`. I suspect it's [this line](https://github.com/SixLabors/Shapes/blob/c38e6d7bdef4ccee7a3d5fdfd7f8d0fb75f86484/src/SixLabors.Shapes/InternalPath.cs#L218).

@tocsoft I think you can solve internally without using heap objects :)

Some usefult info on call times, if you want to speed up things a bit:
![image](https://cloud.githubusercontent.com/assets/6835152/24071295/f7445b32-0bce-11e7-92c8-aa076633b199.png)
I think applying `Vector2` SIMD operations and some inlining in intersection calculator could help here ;)

I don't think we should bother with optimization right in this PR, but I'm adding my comments for information.