@tocsoft 2017-03-18
awesome 👍 look at my code its not even needed at all anymore. Should be a super easy win 😁