@antonfirsov 2017-03-18
Personally I would like to see this merged ASAP, so I can resolve conflicts with changes I made on the `speedup-resize` branch. I can't see any fundamental problem here, we can do optimization later.