@tocsoft 2017-03-18
OK, so everything's green on the build server and I've addressed all the immediate issues (more optimisations can come later) so I'm merging this to unblock @antonfirsov work.